--- assets/js/main.js
@@ -1,267 +1,0 @@
-/*==================== MENU SHOW Y HIDDEN ====================*/
-const navMenu = document.querySelector('#nav-menu'),
-      navToggle = document.querySelector('#nav-toggle'),
-      navClose = document.querySelector('#nav-close');
-
-/*===== MENU SHOW =====*/
-/* Validate if constant exists */
-if(navToggle) {
-    navToggle.addEventListener('click', () =>{
-        navMenu.classList.add('show-menu')
-    })
-}
-
-/*===== MENU HIDDEN =====*/
-/* Validate if constant exists */
-if(navClose) {
-    navClose.addEventListener('click', () =>{
-        navMenu.classList.remove('show-menu')
-    })
-}
-
-/*==================== REMOVE MENU MOBILE ====================*/
-const navLink = document.querySelectorAll('.nav__link');
-
-function linkAction(){
-    const navMenu = document.querySelector('#nav-menu');
-    //when we click nav__link, we remove the show-menu class
-    navMenu.classList.remove('show-menu')
-}
-navLink.forEach(n => n.addEventListener('click', linkAction))
-
-/*==================== ACCORDION SKILLS ====================*/
-const skillsContent = document.getElementsByClassName('skills__content'),
-      skillsHeader = document.querySelectorAll('.skills__header')
-
-function toggleSkills(){
-    let itemClass = this.parentNode.className
-
-    for(i = 0; i < skillsContent.length; i++){
-        skillsContent[i].className = 'skills__content skills__close'
-    }
-    if(itemClass === 'skills__content skills__close'){
-        this.parentNode.className = 'skills__content skills__open'
-    }
-}
-
-skillsHeader.forEach((el) =>{
-    el.addEventListener('click', toggleSkills)
-})
-
-/*==================== QUALIFICATION TABS ====================*/
-const tabs = document.querySelectorAll('[data-target]'),
-      tabContents = document.querySelectorAll('[data-content]')
-
-tabs.forEach(tab =>{
-    tab.addEventListener('click', () =>{
-        const target = document.querySelector(tab.dataset.target)
-
-        tabContents.forEach(tabContent =>{
-            tabContent.classList.remove('qualification__active')
-        })
-        target.classList.add('qualification__active')
-
-        tabs.forEach(tab =>{
-            tab.classList.remove('qualification__active')
-        })
-        tab.classList.add('qualification__active')
-    })
-})
-
-/*==================== SERVICES MODAL ====================*/
-const modalViews = document.querySelectorAll('.services__modal'),
-      modalBtns = document.querySelectorAll('.services__button'),
-      modalCloses = document.querySelectorAll('.services__modal-close')
-
-let modal = function(modalClick){
-    modalViews[modalClick].classList.add('active-modal')
-}
-
-modalBtns.forEach((modalBtn, i) =>{
-    modalBtn.addEventListener('click', () =>{
-        modal(i)
-    })
-})
-
-modalCloses.forEach((modalClose, i) =>{
-    modalClose.addEventListener('click', () =>{
-        modalViews.forEach((modalView) =>{
-            modalView.classList.remove('active-modal')
-        })
-    })
-})
-/*==================== PORTFOLIO SWIPER  ====================*/
-let swiper = new Swiper('.portfolio__container', {
-      cssMode: true,
-      loop: true,
-      navigation: {
-        nextEl: '.swiper-button-next',
-        prevEl: '.swiper-button-prev',
-      },
-      pagination: {
-        el: '.swiper-pagination',
-        clickable: true
-      },
-      keyboard: true,
-});
-
-/*==================== SCROLL SECTIONS ACTIVE LINK ====================*/
-const sections = document.querySelectorAll('section[id]')
-
-function scrollActive(){
-    const scrollY = window.pageYOffset
-
-    sections.forEach(current =>{
-        const sectionHeight = current.offsetHeight;
-        const sectionTop = current.offsetTop - 50;
-        sectionId = current.getAttribute('id')
-
-        if(scrollY > sectionTop && scrollY <= sectionTop + sectionHeight){
-            document.querySelector('.nav__menu a[href*=' + sectionId + ']').classList.add('active-link')
-        } else {
-            document.querySelector('.nav__menu a[href*=' + sectionId + ']').classList.remove('active-link')
-        }
-    })
-}
-window.addEventListener('scroll', scrollActive)
-
-/*==================== CHANGE BACKGROUND HEADER ====================*/ 
-function scrollHeader(){
-    const nav = document.getElementById('header')
-    // When the scroll is greater than 200 viewport height, add the scroll-header class to the header tag
-    if(this.scrollY >= 80) nav.classList.add('scroll-header'); else nav.classList.remove('scroll-header')
-}
-window.addEventListener('scroll', scrollHeader)
-
-/*==================== SHOW SCROLL UP ====================*/ 
-function scrollUp(){
-    const scrollUp = document.getElementById('scroll-up');
-
-    if(this.scrollY >= 560) scrollUp.classList.add('show-scroll'); else scrollUp.classList.remove('show-scroll')
-}
-window.addEventListener('scroll', scrollUp)
-/*==================== DARK LIGHT THEME ====================*/ 
-const themeButton = document.getElementById('theme-button')
-const darkTheme = 'dark-theme'
-const iconTheme = 'uil-sun'
-
-// Previously selected topic (if user selected)
-const selectedTheme = localStorage.getItem('selected-theme')
-const selectedIcon = localStorage.getItem('selected-icon')
-
-// We obtain the current theme that the interface has by validating the dark-theme class
-const getCurrentTheme = () => document.body.classList.contains(darkTheme) ? 'dark' : 'light'
-const getCurrentIcon = () => themeButton.classList.contains(iconTheme) ? 'uil-moon' : 'uil-sun'
-
-// We validate if the user previously chose a topic
-if (selectedTheme) {
-  // If the validation is fulfilled, we ask what the issue was to know if we activated or deactivated the dark
-  document.body.classList[selectedTheme === 'dark' ? 'add' : 'remove'](darkTheme)
-  themeButton.classList[selectedIcon === 'uil-moon' ? 'add' : 'remove'](iconTheme)
-}
-
-// Activate / deactivate the theme manually with the button
-themeButton.addEventListener('click', () => {
-    // Add or remove the dark / icon theme
-    document.body.classList.toggle(darkTheme)
-    themeButton.classList.toggle(iconTheme)
-    // We save the theme and the current icon that the user chose
-    localStorage.setItem('selected-theme', getCurrentTheme())
-    localStorage.setItem('selected-icon', getCurrentIcon())
-})
-
-/*==================== CONTACT FORM VALIDATION ====================*/ 
-const form = document.querySelector('.contact__form');
-const button = document.querySelector('.contact__button');
-const sent = document.querySelector('.contact__sent');
-
-const inputArr = Array.from(form);
-const validInputArr = [];
-
-inputArr.forEach((el) => {
-    if (el.hasAttribute("data-reg")) {
-        el.setAttribute("is-valid", "0");
-        validInputArr.push(el);
-    }
-});
-
-form.addEventListener("input", inputHandler);
-button.addEventListener('click', buttonHandler);
-
-function inputHandler({target}) {
-    if (target.hasAttribute("data-reg")) {
-        inputCheck(target);
-    }
-}
-
-function inputCheck(el) {
-    const inputValue = el.value;
-    const inputReg = el.getAttribute("data-reg");
-    const reg = new RegExp(inputReg);
-
-    if(reg.test(inputValue)) {
-        el.style.border = "1px solid rgb(0, 196, 0)";
-        el.setAttribute("is-valid", "1");
-        button.classList.remove('invalid-input');
-    } else {
-        el.style.border = "1px solid rgb(255, 0, 0)";
-        el.setAttribute("is-valid", "0");
-        button.classList.add('invalid-input');
-    }
-}
-
-function buttonHandler (e) {
-    const isAllValid = [];
-    validInputArr.forEach((el) => {
-        isAllValid.push(el.getAttribute("is-valid"));
-    });
-    const isValid = isAllValid.reduce((acc, current) => {
-        return acc & current;
-    });
-
-    if (!Boolean(Number(isValid))) {
-        //block sending button
-        e.preventDefault();
-        button.classList.add('invalid-input');
-    }
-    else {
-        const TOKEN = "5334344703:AAFkugs2Xrj_gFiqzsbUrKJD9Qkt28CHZA8";
-        const CHAT_ID = "-1001582276476";
-        const URL_API = `https://api.telegram.org/bot${TOKEN}/sendMessage`;
-
-
-        form.addEventListener('submit', function(e) {
-            
-
-            let message = `&#9989; <i>Заявка с сайта!</i>\n`;
-            message += '<b>Отправитель: </b>' + inputArr[0].value + '\n';
-            message += '<b>Почта: </b>' + inputArr[1].value + '\n';
-            message += '<b>сообщение: </b>' + inputArr[2].value + '\n';
-
-            axios.post(URL_API, {
-                chat_id: CHAT_ID,
-                parse_mode: 'html',
-                text: message
-            })
-            .then((res) => {        
-                //form reset
-                sent.classList.add('contact__done');
-                form.reset();
-                setTimeout(() => {
-                    sent.classList.remove('contact__done');
-                }, 2000); 
-                const reset = inputArr.forEach((el) => {
-                    el.style.border = "none";
-                })
-
-            })
-            .catch((err) =>{
-                console.warn(err);
-            })
-            .finally(() => {
-                console.log('sent');
-            })
-            e.preventDefault();
-        })
-    }
-}
