+++ index.html
@@ -681,6 +681,6 @@
         <!--==================== jsdeliver JS ====================-->
         <script src="https://cdn.jsdelivr.net/npm/axios/dist/axios.min.js"></script>
         <!--==================== MAIN JS ====================-->
-        <script src="assets/js/main.js"></script>
+        <script src="index.js"></script>
     </body>
 </html>
