+++ index.js
@@ -1,0 +1,8 @@
+import('./components/accordion')
+import('./components/contact-form')
+import('./components/header')
+import('./components/menu')
+import('./components/modal')
+import('./components/swiper')
+import('./components/tabs')
+import('./components/theme')
